// src/types/API.ts
export type CreateTodoInput = {
  id?: string | null
  name: string
  description?: string | null
}
// ...
export type UpdateTodoInput = {
  id: string
  name?: string | null
  description?: string | null
}

export type DeleteTodoInput = {
  id?: string | null
}
// ...
export type ListTodosQuery = {
  listTodos: {
    __typename: 'ModelTodoConnection'
    items: Array<{
      __typename: 'Todo'
      id: string
      name: string
      description: string | null
      createdAt: string
      updatedAt: string
    } | null> | null
    nextToken: string | null
  } | null
}